Add name field to the class
Since esptool v4.5 the name field is evaluated by the ESPLoader class
in loader.py to construct a reset strategy. If this name starts with
"rfc2217:" the behavior is as earlier. Otherwise the reset is attempted
using direct fcntl access to the underlying serial port via the fileno()
function (which is currently not implemented by esptool-ftdi.py).
Keeping full control over the serial port from within esptool-ftdi.py
seems the much better option so make sure the name starts with
"rfc2217:".